fix/utils refactor
