+++ frontend/src/Chat/components/ChatList/CreateChatRoomModal.tsx
@@ -1,6 +1,6 @@
 import { type ReactElement } from 'react'
 import { ChatRoomSetupModal } from '../utils/Modal/ChatRoomSetupModal'
-import { createChatRoomRequest } from '../utils/requestUtils'
+import { createChatRoomRequest } from '../../../utils/chatRoomAxios'
 
 export const CreateChatRoomModal = (props: {
   user: User
