+++ frontend/src/Chat/components/ChatRoom/BanButton.tsx
@@ -1,5 +1,5 @@
 import { Button } from 'react-bootstrap'
-import { updateChatRoomMembersRequest } from '../utils/requestUtils'
+import { updateChatRoomMembersRequest } from '../../../utils/chatRoomMemberAxios'
 import { isOwner, isTargetBanned } from '../utils/userStatusUtils'
 
 const BanMemberButton = (props: {
