+++ frontend/src/utils/userAxios.tsx
@@ -1,0 +1,30 @@
+import axios from 'axios'
+
+axios.defaults.baseURL = process.env.REACT_APP_BACKEND_HTTP_BASE_URL
+
+export function getAllUsersRequest(callback: (users: User[]) => void): void {
+  axios
+    .get<User[]>('/user/users')
+    .then((response) => {
+      callback(response.data)
+    })
+    .catch((reason) => {
+      alert('エラーです！')
+      console.log(reason)
+    })
+}
+
+export function getUserRequest(
+  userId: number,
+  callback: (user: User) => void
+): void {
+  axios
+    .get<User>('/user/' + String(userId))
+    .then((response) => {
+      callback(response.data)
+    })
+    .catch((reason) => {
+      alert('エラーです！')
+      console.log(reason)
+    })
+}
